feat: 그룹 탈퇴
