+++ src/app/(afterLogin)/group/detail/[groupId]/_lib/fetchLeaveGroup.ts
@@ -1,0 +1,13 @@
+import { fetchWithAuth } from "@/app/(afterLogin)/_lib/FetchWithAuth";
+
+export async function fetchLeaveGroup(groupId: number) {
+  const response = await fetchWithAuth(
+    `/group-participants/groups/${groupId}`,
+    {
+      method: 'DELETE',
+      credentials: 'include',
+      cache: 'no-cache',
+    },
+  );
+  return response;
+}
